refactor: minor types fix
