+++ models/User.ts
@@ -1,8 +1,23 @@
 import type { RowDataPacket } from 'mysql2';
 
-export default interface User {
+export interface User {
   id: number;
-  google_id: string;
+  google_id?: string;
+  email?: string;
+  display_picture?: string;
+  name?: string;
+}
+
+export interface CreateUser {
+  google_id?: string;
+  email?: string;
+  display_picture?: string;
+  name?: string;
+}
+
+// For updating users (all fields optional)
+export interface UpdateUser {
+  google_id?: string;
   email?: string;
   display_picture?: string;
   name?: string;
